refactor: extract app module tests
